requirements-osint.txt
shodan>=1.31.0
networkx>=3.3
python-dotenv>=1.0.0
jinja2>=3.1.4
pytest>=8.2.0
